Extract Prompt Studio serialization backup helpers
